+++ framework/Source/CPTXYPlotSpace.h
@@ -1,9 +1,10 @@
+#import "CPTAnimation.h"
 #import "CPTDefinitions.h"
 #import "CPTPlotSpace.h"
 
 @class CPTPlotRange;
 
-@interface CPTXYPlotSpace : CPTPlotSpace {
+@interface CPTXYPlotSpace : CPTPlotSpace<CPTAnimationDelegate> {
     @private
     CPTPlotRange *xRange;
     CPTPlotRange *yRange;
@@ -17,8 +18,6 @@
     NSTimeInterval lastDeltaTime;
     BOOL isDragging;
     BOOL allowsMomentum;
-    BOOL elasticGlobalXRange;
-    BOOL elasticGlobalYRange;
     NSMutableArray *animations;
 }
 
@@ -30,7 +29,5 @@
 @property (nonatomic, readwrite, assign) CPTScaleType yScaleType;
 
 @property (nonatomic, readwrite) BOOL allowsMomentum;
-@property (nonatomic, readwrite) BOOL elasticGlobalXRange;
-@property (nonatomic, readwrite) BOOL elasticGlobalYRange;
 
 @end